Add nav graph start destination and fix some bugs

+++ app/src/main/java/com/example/android/releasedatekt/domain/Models.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 data class Movie(
     val id: Int,
-    val popularity: Int,
+    val popularity: Float,
     val voteCount: Int,
     val posterPath: String,
     val language: String,
